## Purpose
<!-- Describe the intention of the changes being proposed. What problem
does it solve or functionality does it add? -->
* ...

## Does this introduce a breaking change?
<!-- Mark one with an "x". -->
```
[ ] Yes
[ ] No
```

## Pull Request Type
What kind of change does this Pull Request introduce?

<!-- Please check the one that applies to this PR using "x". -->
```
[ ] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Documentation content changes
[ ] Other... Please describe:
```

## How to Test
*  Get the code

```
git clone [repo-address]
cd [repo-name]
git checkout [branch-name]
npm install
```

* Test the code
<!-- Add steps to run the tests suite and/or manually test -->
```
```

## What to Check
Verify that the following are valid
* ...

## Other Information
<!-- Add any other helpful information that may be needed here. -->

---------

+++ app/backend/Services/AzureBlobStorageService.cs
@@ -17,35 +17,56 @@ internal sealed class AzureBlobStorageService(BlobContainerClient container)
 
                 await using var stream = file.OpenReadStream();
 
-                using var documents = PdfReader.Open(stream, PdfDocumentOpenMode.Import);
-                for (int i = 0; i < documents.PageCount; i++)
+                // if file is an image (end with .png, .jpg, .jpeg, .gif), upload it to blob storage
+                if (Path.GetExtension(fileName).ToLower() is ".png" or ".jpg" or ".jpeg" or ".gif")
                 {
-                    var documentName = BlobNameFromFilePage(fileName, i);
-                    var blobClient = container.GetBlobClient(documentName);
+                    var blobName = BlobNameFromFilePage(fileName);
+                    var blobClient = container.GetBlobClient(blobName);
                     if (await blobClient.ExistsAsync(cancellationToken))
                     {
                         continue;
                     }
 
-                    var tempFileName = Path.GetTempFileName();
-
-                    try
+                    var url = blobClient.Uri.AbsoluteUri;
+                    await using var fileStream = file.OpenReadStream();
+                    await blobClient.UploadAsync(fileStream, new BlobHttpHeaders
                     {
-                        using var document = new PdfDocument();
-                        document.AddPage(documents.Pages[i]);
-                        document.Save(tempFileName);
-
-                        await using var tempStream = File.OpenRead(tempFileName);
-                        await blobClient.UploadAsync(tempStream, new BlobHttpHeaders
+                        ContentType = "image"
+                    }, cancellationToken: cancellationToken);
+                    uploadedFiles.Add(blobName);
+                }
+                else if (Path.GetExtension(fileName).ToLower() is ".pdf")
+                {
+                    using var documents = PdfReader.Open(stream, PdfDocumentOpenMode.Import);
+                    for (int i = 0; i < documents.PageCount; i++)
+                    {
+                        var documentName = BlobNameFromFilePage(fileName, i);
+                        var blobClient = container.GetBlobClient(documentName);
+                        if (await blobClient.ExistsAsync(cancellationToken))
                         {
-                            ContentType = "application/pdf"
-                        }, cancellationToken: cancellationToken);
+                            continue;
+                        }
 
-                        uploadedFiles.Add(documentName);
-                    }
-                    finally
-                    {
-                        File.Delete(tempFileName);
+                        var tempFileName = Path.GetTempFileName();
+
+                        try
+                        {
+                            using var document = new PdfDocument();
+                            document.AddPage(documents.Pages[i]);
+                            document.Save(tempFileName);
+
+                            await using var tempStream = File.OpenRead(tempFileName);
+                            await blobClient.UploadAsync(tempStream, new BlobHttpHeaders
+                            {
+                                ContentType = "application/pdf"
+                            }, cancellationToken: cancellationToken);
+
+                            uploadedFiles.Add(documentName);
+                        }
+                        finally
+                        {
+                            File.Delete(tempFileName);
+                        }
                     }
                 }
             }
@@ -53,7 +74,7 @@ internal sealed class AzureBlobStorageService(BlobContainerClient container)
             if (uploadedFiles.Count is 0)
             {
                 return UploadDocumentsResponse.FromError("""
-                    No files were uploaded. Either the files already exist or the files are not PDFs.
+                    No files were uploaded. Either the files already exist or the files are not PDFs or images.
                     """);
             }
